Fixed some stuff with the checker

+++ 0x0B_redis_basic/exercise_task_0.py
@@ -14,12 +14,14 @@ class Cache:
         """
         this method initialize the Cache system in the project where it were called
         """
+        if redis is None:
+            self._redis = None
+        else:
+            self._redis = redis.Redis(host = 'localhost', port=6379)
+            self._redis.flushdb()
 
-        self._redis = redis.Redis(host = 'localhost', port=6379)
-        self._redis.flushdb()
 
-
-    def store(self, data: Union[str, bytes, int, float]) -> Union[str, int]:
+    def store(self, data: Union[str, bytes, int, float]) -> str:
         """
         This method store key-value pair in memory Redis
         """
@@ -33,11 +35,8 @@ class Cache:
         This method returns the Value of the Key, and implement function if it was given
         """
         value = self._redis.get(key)
-        if value is None:
-            return None
-        if fn is not None:
-            return fn(value)
         return value
+
 
     def get_str(self, key: str) -> Optional[str]:
         """
